move outcomes to a separate file
