+++ grug/types/src/app.rs
@@ -1,16 +1,10 @@
 use {
-    crate::{
-        Addr, CommitmentStatus, Duration, Event, EventStatus, EvtAuthenticate, EvtBackrun, EvtCron,
-        EvtFinalize, EvtWithhold, GenericResult, Hash256, Json, Message, Timestamp,
-    },
+    crate::{Addr, Duration, Hash256, Json, Message, Timestamp},
     borsh::{BorshDeserialize, BorshSerialize},
     hex_literal::hex,
     serde::{Deserialize, Serialize},
     serde_with::skip_serializing_none,
-    std::{
-        collections::{BTreeMap, BTreeSet},
-        fmt::{self, Display},
-    },
+    std::collections::{BTreeMap, BTreeSet},
 };
 
 /// The mock up sender address used for executing genesis messages.
@@ -110,167 +104,4 @@ pub struct ContractInfo {
     pub code_hash: Hash256,
     pub label: Option<String>,
     pub admin: Option<Addr>,
-}
-
-/// Outcome of performing an operation that is not a full tx. These include:
-///
-/// - processing a message;
-/// - executing a cronjob;
-/// - performing a `CheckTx` call.
-///
-/// Includes the events emitted, and gas consumption.
-#[derive(Serialize, Deserialize, BorshSerialize, BorshDeserialize, Debug, Clone, PartialEq, Eq)]
-#[serde(deny_unknown_fields)]
-#[must_use = "`Outcome` must be checked for success or error with `should_succeed`, `should_fail`, or similar methods."]
-pub struct CheckTxOutcome {
-    // `None` means the call was done with unlimited gas, such as cronjobs.
-    pub gas_limit: Option<u64>,
-    pub gas_used: u64,
-    pub result: GenericResult<()>,
-}
-
-#[derive(Serialize, Deserialize, BorshSerialize, BorshDeserialize, Debug, Clone, PartialEq, Eq)]
-#[serde(deny_unknown_fields)]
-#[must_use = "`Outcome` must be checked for success or error with `should_succeed`, `should_fail`, or similar methods."]
-pub struct CronOutcome {
-    // `None` means the call was done with unlimited gas, such as cronjobs.
-    pub gas_limit: Option<u64>,
-    pub gas_used: u64,
-    pub cron_event: CommitmentStatus<EvtCron>,
-}
-
-impl CronOutcome {
-    pub fn new(
-        gas_limit: Option<u64>,
-        gas_used: u64,
-        cron_event: CommitmentStatus<EvtCron>,
-    ) -> Self {
-        Self {
-            gas_limit,
-            gas_used,
-            cron_event,
-        }
-    }
-}
-/// Outcome of processing a transaction.
-///
-/// Different from `Outcome`, which can either succeed or fail, a transaction
-/// can partially succeed. A typical such scenario is:
-///
-/// - `withhold_fee` succeeds
-/// - `authenticate` succeeds,
-/// - one of the messages fail
-/// - `finalize_fee` succeeds
-///
-/// In this case, state changes from fee handling (e.g. deducting the fee from
-/// the sender account) and authentication (e.g. incrementing the sender account's
-/// sequence number) will be committed, and relevant events emitted to reflect
-/// this. However, state changes and events from the messages are discarded.
-#[derive(Serialize, Deserialize, BorshSerialize, BorshDeserialize, Debug, Clone, PartialEq, Eq)]
-#[serde(deny_unknown_fields)]
-#[must_use = "`TxOutcome` must be checked for success or error with `should_succeed`, `should_fail`, or similar methods."]
-pub struct TxOutcome {
-    pub gas_limit: u64,
-    pub gas_used: u64,
-    pub result: GenericResult<()>,
-    pub events: TxEvents,
-}
-
-#[derive(Debug, PartialEq, Eq)]
-pub struct TxSuccess {
-    pub gas_limit: u64,
-    pub gas_used: u64,
-    pub events: TxEvents,
-}
-
-#[derive(Debug, PartialEq, Eq)]
-pub struct TxError {
-    pub gas_limit: u64,
-    pub gas_used: u64,
-    pub error: String,
-    pub events: TxEvents,
-}
-
-// `TxError` must implement `ToString`, such that it satisfies that trait bound
-// required by `ResultExt::should_fail_with_error`.
-impl Display for TxError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{self:?}",)
-    }
-}
-
-#[derive(Serialize, Deserialize, BorshSerialize, BorshDeserialize, Debug, Clone, PartialEq, Eq)]
-pub struct TxEvents {
-    pub withhold: CommitmentStatus<EvtWithhold>,
-    pub authenticate: CommitmentStatus<EvtAuthenticate>,
-    pub msgs_and_backrun: CommitmentStatus<MsgsAndBackrunEvents>,
-    pub finalize: CommitmentStatus<EvtFinalize>,
-}
-
-impl TxEvents {
-    pub fn new(withhold: CommitmentStatus<EvtWithhold>) -> Self {
-        Self {
-            withhold,
-            authenticate: CommitmentStatus::NotReached,
-            msgs_and_backrun: CommitmentStatus::NotReached,
-            finalize: CommitmentStatus::NotReached,
-        }
-    }
-
-    pub fn finalize_fails(self, finalize: CommitmentStatus<EvtFinalize>, cause: &str) -> Self {
-        fn update<T>(evt: CommitmentStatus<T>, cause: &str) -> CommitmentStatus<T> {
-            if let CommitmentStatus::Committed(event) = evt {
-                CommitmentStatus::Reverted {
-                    event,
-                    revert_by: cause.to_string(),
-                }
-            } else {
-                evt
-            }
-        }
-
-        TxEvents {
-            withhold: update(self.withhold, cause),
-            authenticate: update(self.authenticate, cause),
-            msgs_and_backrun: update(self.msgs_and_backrun, cause),
-            finalize,
-        }
-    }
-}
-
-#[derive(Serialize, Deserialize, BorshSerialize, BorshDeserialize, Debug, Clone, PartialEq, Eq)]
-pub struct MsgsAndBackrunEvents {
-    pub msgs: Vec<EventStatus<Event>>,
-    pub backrun: EventStatus<EvtBackrun>,
-}
-
-impl MsgsAndBackrunEvents {
-    pub fn base() -> Self {
-        Self {
-            msgs: vec![],
-            backrun: EventStatus::NotReached,
-        }
-    }
-}
-
-impl Default for TxEvents {
-    fn default() -> Self {
-        Self {
-            withhold: CommitmentStatus::NotReached,
-            authenticate: CommitmentStatus::NotReached,
-            msgs_and_backrun: CommitmentStatus::NotReached,
-            finalize: CommitmentStatus::NotReached,
-        }
-    }
-}
-
-#[derive(Debug)]
-/// Outcome of executing a block.
-pub struct BlockOutcome {
-    /// The Merkle root hash after executing this block.
-    pub app_hash: Hash256,
-    /// Results of executing the cronjobs.
-    pub cron_outcomes: Vec<CronOutcome>,
-    /// Results of executing the transactions.
-    pub tx_outcomes: Vec<TxOutcome>,
 }
